+++ activerecord/lib/active_record/connection_adapters/abstract/schema_creation.rb
@@ -100,7 +100,12 @@ module ActiveRecord
         end
 
         def add_column_options!(sql, options)
-          sql << " DEFAULT #{quote_default_expression(options[:default], options[:column])}" if options_include_default?(options)
+          if options_include_default?(options)
+            if options[:null] == true
+              sql << " NULL"
+            end
+            sql << " DEFAULT #{quote_default_expression(options[:default], options[:column])}"
+          end
           # must explicitly check for :null to allow change_column to work on migrations
           if options[:null] == false
             sql << " NOT NULL"